When reading migrations table, use Quorum consistency
Use lower consitency to read the migrations table, in order to figure
out if the current migration needs to be run or not. Other operations
that will actually insert data or change schemas, will still use ALL
consistency.

This will allow successfully running the migration code in situations
where NO schema changes are required and ALL the nodes are not
avaialble.